Unit tests for create contact
