+++ customer-os-api/server/graph/resolver/test_queries/create_tenant_user.txt
@@ -1,0 +1,9 @@
+mutation {
+  createTenantUser(input: {firstName:"first", lastName:"last", email:"user@openline.ai"}) {
+      id
+      firstName
+      lastName
+      email
+      createdAt
+    }
+}
